requirements.txt
python-dotenv
pymongo[srv]
requests
beautifulsoup4
lxml
openai
googlemaps
asyncio